update diary tabs
